REF: test_nlargest
